+++ packages/telemetry/test/helper.js
@@ -1,0 +1,23 @@
+const fastify = require('fastify')
+const telemetryPlugin = require('../lib/telemetry')
+
+async function setupApp (pluginOpts, routeHandler, teardown) {
+  const app = fastify()
+  await app.register(telemetryPlugin, pluginOpts)
+  app.get('/test', routeHandler)
+  app.ready()
+  teardown(async () => {
+    await app.close()
+    const { exporters } = app.openTelemetry
+    exporters.forEach(exporter => {
+      if (exporter.constructor.name === 'InMemorySpanExporter') {
+        exporter.reset()
+      }
+    })
+  })
+  return app
+}
+
+module.exports = {
+  setupApp
+}
